Voyage page moved notice, calculator looks for event start date
The `voyage.tsx` page and `voyagecalculator.tsx` component have been renamed with a `_legacy` postfix. Keeping them around until it's safe to delete them.

Now in the `voyage.tsx` page is a notice to users to update their bookmarks to the new player tools landing page (/playertools). It does not automatically redirect them at this time.

TODO: `profile.tsx` and `fleet_info.tsx` refer to the old /voyage link. Those should be changed to the correct /playertools link after this is merged.

The voyage calculator now considers the event start date when pre-excluding crew. I had to re-arrange some checks because the suggested change sometimes yielded an empty event list when previously the code was always expecting at least 1 event.
